k_moneyball/pyproject.toml
[tool.poetry]
name = "k-moneyball"
version = "0.1.0"
description = ""
authors = ["MoonSeoJun <msj00130@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
pandas = "^2.1.1"
psycopg2-binary = "^2.9.9"
pymongo = "^4.5.0"
kafka-python = "^2.0.2"
apache-airflow = {extras = ["cncf-kubernetes"], version = "^2.7.2"}
connexion = "^2.14.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true